refactor out into wb link helper
