+++ OpenHD/ohd_interface/src/wb_link_helper.cpp
@@ -125,3 +125,12 @@ std::vector<std::string> openhd::wb::get_card_names(const std::vector<WiFiCard>&
   }
   return ret;
 }
+
+bool openhd::wb::has_any_rtl8812au(const std::vector<WiFiCard>& cards) {
+  for(const auto& card: m_broadcast_cards){
+    if(card.type==WiFiCardType::Realtek8812au){
+      return true;
+    }
+  }
+  return false;
+}
